+++ Libplanet.Tests/Store/TrieStateStoreCommitTest.cs
@@ -1,9 +1,11 @@
+using System;
 using System.Linq;
 using System.Security.Cryptography;
 using Bencodex.Types;
 using Libplanet.Common;
 using Libplanet.Store;
 using Libplanet.Store.Trie;
+using Libplanet.Store.Trie.Nodes;
 using Xunit;
 
 namespace Libplanet.Tests.Store
@@ -20,6 +22,7 @@ namespace Libplanet.Tests.Store
 
             Assert.Null(emptyTrie.Root);
             Assert.True(stateStore.GetStateRoot(emptyRootHash).Recorded);
+            Assert.Null(stateStore.GetStateRoot(emptyRootHash).Root);
             Assert.False(keyValueStore.Exists(new KeyBytes(emptyRootHash.ByteArray)));
 
             emptyTrie = stateStore.Commit(emptyTrie);
@@ -56,6 +59,20 @@ namespace Libplanet.Tests.Store
             Assert.Equal(2, trie.IterateValues().Count());
             Assert.Equal(new Text("2c73"), trie.Get(new KeyBytes(new byte[] { 0x2c, 0x73 })));
             Assert.Equal(new Text("234f"), trie.Get(new KeyBytes(new byte[] { 0x23, 0x4f })));
+        }
+
+        [Fact]
+        public void CommittedNonEmptyTrieRootIsHashNode()
+        {
+            IKeyValueStore keyValueStore = new MemoryKeyValueStore();
+            IStateStore stateStore = new TrieStateStore(keyValueStore);
+            ITrie trie = stateStore.GetStateRoot(null);
+            trie = trie.Set(new KeyBytes(Array.Empty<byte>()), new Integer(1));
+            trie = stateStore.Commit(trie);
+            HashNode root = Assert.IsType<HashNode>(trie.Root);
+            trie = stateStore.GetStateRoot(trie.Hash);
+            Assert.IsType<HashNode>(trie.Root);
+            Assert.Equal(root, trie.Root);
         }
     }
 }